# api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
  useWebSearch?: boolean;
}

export interface ChatResponse {
  reply: string;
  sources?: string[];
  solutionId?: number;
}

export interface StoreSolutionRequest {
  problemKeywords: string;
  problemSummary: string;
  solution: string;
}

export interface StoreSolutionResponse {
  id: number;
}