# openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chatbot
    description: AI chatbot operations
paths:
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chatbot]
      summary: Send a message to the AI chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /chat/solution:
    post:
      operationId: storeChatSolution
      tags: [chatbot]
      summary: Store a helpful solution for future reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreSolutionRequest"
      responses:
        "201":
          description: Solution stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSolutionResponse"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      required:
        - role
        - content
    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        useWebSearch:
          type: boolean
      required:
        - messages
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        sources:
          type: array
          items:
            type: string
        solutionId:
          type: integer
      required:
        - reply
    StoreSolutionRequest:
      type: object
      properties:
        problemKeywords:
          type: string
        problemSummary:
          type: string
        solution:
          type: string
      required:
        - problemKeywords
        - problemSummary
        - solution
    StoreSolutionResponse:
      type: object
      properties:
        id:
          type: integer
      required:
        - id